Addded separate csv to json class to separate logic from UI
