+++ src/screens/authenticated/merlin/sub-components/mailing/index.tsx
@@ -4,6 +4,8 @@ import './styles.css'
 import Papa from 'papaparse';
 import CSVReader from './csv-reader';
 import { counter } from '@fortawesome/fontawesome-svg-core';
+import { CSVtoJSON } from '../../../../../services/CSVtoJSON/csv-to-json-service'
+
 
 type CSVFILE = {
     data: []
@@ -14,34 +16,13 @@ const MerlinMailing = () => {
     const [csvLength, updateCSVLength] = useState<number | null>(null);
     const str = JSON.stringify(csvFile)
 
-    const handleCSVtoJSON = () =>{
-        const obj = {}
-      const {data} = csvFile;
-      const length = Object.keys(data).length
+    const CSVService = new CSVtoJSON;
 
-      var x = 1
-      var headerObject:any = {}
-
-      data.forEach((colCounter: any[])=>{
-        if (x === 1){
-            for (var headerCounter=0; headerCounter< colCounter.length; headerCounter++){
-
-                console.log(colCounter[headerCounter]);
-                const currentHeader:string = colCounter[headerCounter]
-                const tempVar = currentHeader
-                console.log(currentHeader +' oten')
-
-               headerObject[tempVar] = tempVar
-            }
-            console.log(headerObject)
-        } else{
-            console.log(colCounter)
-        }
-        x = x+1
-      })
-      
-  
+    const handleCSV = () => {
+        CSVService.handleCSVtoJSON(csvFile)
     }
+
+    
  
   return (
     <div className='main-div col-md-6'>
@@ -52,7 +33,7 @@ const MerlinMailing = () => {
             <CSVReader csvFile={csvFile} updateCSVFile={updateCSVFile}></CSVReader>
             </Col>
         </Row>
-            <Button className='my-2' onClick={handleCSVtoJSON}>Parse</Button>
+            <Button className='my-2' onClick={()=>handleCSV()}>Parse</Button>
     </Container>
     </div>
   )
